refactor: deck의 책임 간소화 및 Cards의 책임 추가
- 기존 deck에서 52장의 카드를 생성하고 섞는 로직을 Cards로 위임했습니다.
이에 따라, deckTest를 지우고 CardsTest를 생성했습니다.
기존에 있던 CardsTest는 ResultTest로 renmae 했습니다.
