+++ src/test/java/domain/testDeck.java
@@ -1,33 +1,22 @@
 package domain;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class testDeck {
     private final List<Card> cards;
 
-    public testDeck(List<Card> cards){
+    public testDeck(List<Card> cards) {
         this.cards = new ArrayList<>(cards);
     }
 
-    public static testDeck createStandardDeck() {
-        List<Card> cards = new ArrayList<>();
-        for (Shape shape : Shape.values()) {
-            for (Number number : Number.values()) {
-                cards.add(new Card(shape, number));
-            }
-        }
-        return new testDeck(cards);
-    }
-
-    public static testDeck shuffleCards() {
-        List<Card> shuffledCards = new ArrayList<>();
-        Collections.shuffle(shuffledCards);
+    public static testDeck createDeck() {
+        List<Card> standardCards = Cards.createCards();
+        List<Card> shuffledCards = Cards.shuffleCards(standardCards);
         return new testDeck(shuffledCards);
     }
 
-    public static Integer getDeckSize(testDeck testDeck){
+    public static Integer getDeckSize(testDeck testDeck) {
         return testDeck.cards.size();
     }
 
@@ -36,11 +25,5 @@ public class testDeck {
     public Card distributeCard() {
         Card card = cards.removeFirst();
         return card;
-    }
-
-    private List<Card> shuffleCards(List<Card> cards) {
-        List<Card> shuffledCards = new ArrayList<>(cards);
-        Collections.shuffle(shuffledCards);
-        return shuffledCards;
     }
 }
